+++ scripts/deploy-in-vscode.js
@@ -4,7 +4,7 @@
 
 
 // Imports
-import { existsSync, emptyDirSync, mkdirSync, copySync } from 'fs-extra';
+import fs from 'fs-extra';
 
 console.log('---- START DEPLOY IN VSCODE INSTALLED npm-groovy-lint PACKAGE ----');
 
@@ -15,16 +15,16 @@ const targetPath = `${vsCodeGroovyLintPath}/server/node_modules/npm-groovy-lint`
 console.info(`GroovyLint: Starting copying package in vscode for testing`);
 
 // Reset target folder
-if (existsSync(targetPath)) {
-    emptyDirSync(targetPath);
+if (fs.existsSync(targetPath)) {
+    fs.emptyDirSync(targetPath);
 }
 else {
-    mkdirSync(targetPath);
+    fs.mkdirSync(targetPath);
 }
 
 // Copy files into dest folder
 for (const path of ['package.json', 'README.md', 'CHANGELOG.md', 'LICENSE', 'lib']) {
-    copySync(path, `${targetPath}/${path}`);
+    fs.copySync(path, `${targetPath}/${path}`);
 }
 
 console.info(`GroovyLint: Copied files into ${targetPath}`);